refactor: MSW 에러 시나리오 분리
