+++ FE/src/mocks/scenarios.ts
@@ -1,0 +1,5 @@
+import { defaultHandlers } from '@/mocks/handlers/default';
+
+export const scenarios = {
+  default: defaultHandlers,
+};
